Start migrating compiler.js
